feat: updated efun contracts and dot-call support (#318)
* clean: comments about func_spec.c

* feat: add allow_dot_call no_dot_call pragma

* fix: efun token width masking

* fix: vital object protection in destruct_object

* test: enable pedantic mode for unit-tests

* chore: fix msvc warning

* fix: error handling in mudlib_connect

* fix: restore PACKAGE_SOCKETS in options.h

* feat: updated efun contract and allow_dot_call set

+++ lib/lpc/identifier.c
@@ -79,7 +79,7 @@ ident_hash_elem_t *lookup_ident (const char *name) {
  * @param name The name of the identifier. No reference is made to the string after this call.
  * @return A pointer to the identifier hash element.
  */
-ident_hash_elem_t* find_or_add_perm_ident (char *name, short token_flag) {
+ident_hash_elem_t* find_or_add_perm_ident (char *name, uint32_t token_flag) {
   int h = IdentHash (name);
   ident_hash_elem_t *hptr, *hptr2;
